fix: syscalls should return long number

+++ sys_save_recover/sys_save_recover.c
@@ -1,12 +1,12 @@
 #include<linux/linkage.h>
 #include<linux/module.h>
 
-static int saved_number;
+static long saved_number;
 
-asmlinkage int sys_save(int number) {
+asmlinkage long sys_save(long number) {
 	return saved_number = number;
 }
 
-asmlinkage int sys_recover(void) {
+asmlinkage long sys_recover(void) {
 	return saved_number;
 }
